+++ api/pkg/handler/routing.go
@@ -13,8 +13,8 @@ import (
 	"github.com/kubermatic/kubermatic/api/pkg/handler/v1/common"
 	"github.com/kubermatic/kubermatic/api/pkg/provider"
 	"github.com/kubermatic/kubermatic/api/pkg/serviceaccount"
+	"github.com/kubermatic/kubermatic/api/pkg/watcher"
 
-	common2 "github.com/kubermatic/kubermatic/api/pkg/watcher/common"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
 )
@@ -49,7 +49,7 @@ type Routing struct {
 	userInfoGetter              provider.UserInfoGetter
 	settingsProvider            provider.SettingsProvider
 	adminProvider               provider.AdminProvider
-	settingsWatcher             common2.SettingsWatcher
+	settingsWatcher             watcher.SettingsWatcher
 }
 
 // NewRouting creates a new Routing.
@@ -81,7 +81,7 @@ func NewRouting(
 	userInfoGetter provider.UserInfoGetter,
 	settingsProvider provider.SettingsProvider,
 	adminProvider provider.AdminProvider,
-	settingsWatcher common2.SettingsWatcher,
+	settingsWatcher watcher.SettingsWatcher,
 ) Routing {
 	return Routing{
 		log:                         logger,
