+++ api/pkg/watcher/kubernetes/settings.go
@@ -1,4 +1,4 @@
-package watcher
+package kubernetes
 
 import (
 	"fmt"
